Backtest and Logging logic completed
Logic for the backtester is completed for the most part. Small additions will be added as needed. For now it is functional as long as a strategy is available for use.
Logging logic is also completed. Logs can be captured across any module. Three main sources for the logs are the backtester, paper, and live trading which have separate logs for each. All other logs will be placed in trading_all log file.